[feature/paltolaguirre-svc42]
Se cambio el tipo de dato de importe por float64, falta redondear a 4 decimales cuando agrupo las cuentas

+++ actions.go
@@ -35,12 +35,12 @@ type strTransaccionesIdsAsientosContablesManuales struct {
 
 type strCuentaImporte struct {
 	Cuentaid      int     `json:"cuentaid"`
-	Importecuenta float32 `json:"importecuenta"`
+	Importecuenta float64 `json:"importecuenta"`
 }
 
 type strCuentaImporteTipoGrilla struct {
 	Cuentaid      int     `json:"cuentaid"`
-	Importecuenta float32 `json:"importecuenta"`
+	Importecuenta float64 `json:"importecuenta"`
 	Tipogrilla    int     `json:"tipogrilla"`
 }
 
@@ -341,7 +341,7 @@ func LiquidacionRemove(w http.ResponseWriter, r *http.Request) {
 
 func LiquidacionContabilizar(w http.ResponseWriter, r *http.Request) {
 	fmt.Println("La URL accedida: " + r.URL.String())
-	var mapCuentasImportes = make(map[int]float32)
+	var mapCuentasImportes = make(map[int]float64)
 	var strCuentaImporteTipoGrillas []strCuentaImporteTipoGrilla
 	var strCuentasImportes []strCuentaImporte
 	tokenValido, tokenAutenticacion := apiclientautenticacion.CheckTokenValido(w, r)
@@ -406,7 +406,7 @@ func checkLiquidacionesNoContabilizadas(liquidaciones []structLiquidacion.Liquid
 	return len(liquidaciones) == strCheckLiquidacionesNoContabilizadas.Cantidadliquidacionesnocontabilizadas
 }
 
-func generarAsientoManualDesdeMonolitico(w http.ResponseWriter, r *http.Request, liquidaciones []structLiquidacion.Liquidacion, mapCuentasImportes map[int]float32, tokenAutenticacion *structAutenticacion.Security, descripcion string, fechaasiento string, asientomanualtransaccionid int, db *gorm.DB) {
+func generarAsientoManualDesdeMonolitico(w http.ResponseWriter, r *http.Request, liquidaciones []structLiquidacion.Liquidacion, mapCuentasImportes map[int]float64, tokenAutenticacion *structAutenticacion.Security, descripcion string, fechaasiento string, asientomanualtransaccionid int, db *gorm.DB) {
 	var cuentasImportes []monoliticComunication.StrCuentaImporte
 	cuentasImportes = obtenerCuentasImportesLiquidacion(mapCuentasImportes)
 	datosAsientoContableManual := monoliticComunication.Generarasientomanual(w, r, cuentasImportes, tokenAutenticacion, descripcion, fechaasiento)
@@ -545,7 +545,7 @@ func obtenerCuentasImportes(strCuentaImporteTipoGrillas []strCuentaImporteTipoGr
 	}
 }
 
-func agruparCuentas(strCuentasImportes []strCuentaImporte, mapCuentasImportes map[int]float32) {
+func agruparCuentas(strCuentasImportes []strCuentaImporte, mapCuentasImportes map[int]float64) {
 	for i := 0; i < len(strCuentasImportes); i++ {
 		cuentaContable := strCuentasImportes[i].Cuentaid
 		importeUnitario := strCuentasImportes[i].Importecuenta
@@ -555,7 +555,7 @@ func agruparCuentas(strCuentasImportes []strCuentaImporte, mapCuentasImportes ma
 	}
 }
 
-func obtenerCuentasImportesLiquidacion(mapCuentasImportes map[int]float32) []monoliticComunication.StrCuentaImporte {
+func obtenerCuentasImportesLiquidacion(mapCuentasImportes map[int]float64) []monoliticComunication.StrCuentaImporte {
 	var arrayStrCuentaImporte []monoliticComunication.StrCuentaImporte
 
 	for cuenta, importe := range mapCuentasImportes {
